Implementation of basic comparative sorting methods -Selection Sort,Insertion Sort, Shell Sort and Counting Sort.
Problem statement   : Write a generalised function that takes a parameter to indicate the mode(say1 for decreasing order, 2 for increasing order, 3 for increasing order with the Nth element out of order, 4 for a randomly generated element values), to create a list of elements.
The parameter indicating the number of elements in a statistically allocated array (the maximum size is large enough to run possible iterations to test the time complexity, say 10,000,000)will be a multiple of 10.Also write appropriate function to create a copy of thye list and  to display the list contents.
Using above functions ,write a menu driven c program to order the list in
ascending sequence using - the selection sort, the insertion sort,the counting sort and the shell sort.